Puts the demomode launcher above the keyguard; traps keypresses
This should make it so that a keyguard won't ever be shown until an app
is attempted to be launched. Ideally, the device should be set up
without a keyguard so that it's never shown to the user.

This also traps extra keypresses when locked, so as to avoid things like
long-pressing the search button to trigger a voice search.